+++ server/component/style/entryList/EntryListView.php
@@ -53,8 +53,10 @@ class EntryListView extends StyleView
      */
     public function output_content()
     {
-        $this->model->loadChildren($this->model->get_entry_record());
-        $this->set_children($this->model->get_children());
+        if ($this->model->get_user_input()->is_there_user_input_change()) {
+            $this->model->loadChildren($this->model->get_entry_record());
+            $this->set_children($this->model->get_children());
+        }
         if ($this->load_as_table) {
             require __DIR__ . "/tpl_entryList_table.php";            
         } else {