+++ server/component/style/loop/LoopView.php
@@ -52,6 +52,10 @@ class LoopView extends StyleView
      */
     public function output_content()
     {
+        if ($this->model->get_user_input()->is_there_user_input_change()) {
+            $this->model->loadChildren($this->model->get_entry_record());
+            $this->set_children($this->model->get_children());
+        }
         $this->output_children();
     }
 }