+++ server/db/FUN_PRO_VIEWS/34_view_dataTables.sql
@@ -2,6 +2,7 @@ DROP VIEW IF EXISTS view_dataTables;
 CREATE VIEW view_dataTables
 AS
 SELECT id, 
+`name` AS name_id,
 CASE 
 	WHEN IFNULL(displayName, '') = '' THEN `name`
     ELSE displayName