+++ CHANGELOG.md
@@ -10,6 +10,7 @@
  - remove fields `submit_and_send_label`, `email_body`, `submit_and_send_email`, `email_address` and `email_subject` from form styles. This functionality can be achieved with actions
  - remove style `emailForm`
  - add new style `entryRecordDelete`
+ - check if there is an `user_input` change that came from a controller somewhere. It is used to reload the view children for the dynamic styles like `entryList`, `entryRecord`, `dataContainer` and `Loop`
 
 ### Bugfix
  - set values for `JSON` content fields properly in CMS